4.2.0 Release

+++ contracts/v1/assets/ContentAsset.sol
@@ -153,13 +153,11 @@ contract ContentAsset is Named, Versioned, HubDependent, Initializable {
         );
         uint32 r0 = params.r0();
 
-        if ((timeNow < commitPhaseEnd) && (commitsCount < r0)) {
+        if ((timeNow < commitPhaseEnd) && (commitsCount < r0))
             revert ContentAssetErrors.CommitPhaseOngoing(agreementId);
-        } else if ((timeNow < epochEnd) && (commitsCount >= r0)) {
+        else if ((timeNow < epochEnd) && (commitsCount >= r0))
             revert ContentAssetErrors.CommitPhaseSucceeded(agreementId);
-        } else if (timeNow > epochEnd) {
-            revert ContentAssetErrors.FirstEpochHasAlreadyEnded(agreementId);
-        }
+        else if (timeNow > epochEnd) revert ContentAssetErrors.FirstEpochHasAlreadyEnded(agreementId);
 
         uint96 tokenAmount = sasProxy.getAgreementTokenAmount(agreementId);
 
@@ -210,7 +208,7 @@ contract ContentAsset is Named, Versioned, HubDependent, Initializable {
         uss.setUnfinalizedState(tokenId, assertionId);
         uss.setIssuer(tokenId, msg.sender);
 
-        if (!sasProxy.agreementV1U1Exists(agreementId)) {
+        if (!sasProxy.agreementV1U1Exists(agreementId))
             sasProxy.createV1U1ServiceAgreementObject(
                 agreementId,
                 startTime,
@@ -220,7 +218,6 @@ contract ContentAsset is Named, Versioned, HubDependent, Initializable {
                 scoreFunctionIdAndProofWindowOffsetPerc[0],
                 scoreFunctionIdAndProofWindowOffsetPerc[1]
             );
-        }
 
         if (updateTokenAmount != 0) serviceAgreementV1.addUpdateTokens(msg.sender, agreementId, updateTokenAmount);
 
@@ -257,25 +254,22 @@ contract ContentAsset is Named, Versioned, HubDependent, Initializable {
         bytes32 unfinalizedState = uss.getUnfinalizedState(tokenId);
         uint256 unfinalizedStateIndex = cas.getAssertionIdsLength(tokenId);
 
-        if (unfinalizedState == bytes32(0)) {
+        if (unfinalizedState == bytes32(0))
             revert ContentAssetErrors.NoPendingUpdate(contentAssetStorageAddress, tokenId);
-        } else if (
+        else if (
             block.timestamp <=
             sasProxy.getUpdateCommitsDeadline(keccak256(abi.encodePacked(agreementId, unfinalizedStateIndex)))
-        ) {
+        )
             revert ContentAssetErrors.PendingUpdateFinalization(
                 contentAssetStorageAddress,
                 tokenId,
                 unfinalizedStateIndex
             );
-        }
 
         uint96 updateTokenAmount = sasProxy.getAgreementUpdateTokenAmount(agreementId);
-        if (sasProxy.agreementV1Exists(agreementId)) {
-            sasProxy.deleteServiceAgreementV1U1Object(agreementId);
-        } else {
-            sasProxy.setAgreementUpdateTokenAmount(agreementId, 0);
-        }
+        if (sasProxy.agreementV1Exists(agreementId)) sasProxy.deleteServiceAgreementV1U1Object(agreementId);
+        else sasProxy.setAgreementUpdateTokenAmount(agreementId, 0);
+
         sasProxy.transferV1U1AgreementTokens(msg.sender, updateTokenAmount);
 
         uss.deleteIssuer(tokenId);
@@ -316,9 +310,8 @@ contract ContentAsset is Named, Versioned, HubDependent, Initializable {
         uint128 epochLength;
         (startTime, oldEpochsNumber, epochLength, , ) = serviceAgreementStorageProxy.getAgreementData(agreementId);
 
-        if (block.timestamp > startTime + oldEpochsNumber * epochLength) {
+        if (block.timestamp > startTime + oldEpochsNumber * epochLength)
             revert ContentAssetErrors.AssetExpired(tokenId);
-        }
 
         sasV1.extendStoringPeriod(msg.sender, agreementId, epochsNumber, tokenAmount);
